feat: Add new category type "Penklær" and corresponding icons
This commit adds a new category type "Penklær" to the Enums.ts file. It also includes the necessary icons (Shirt, Blouse, Singlet) in the Category.tsx file to represent this category type in the product form. The addition of "Penklær" expands the available options for users when selecting a category for their products.