feat: Update links and accordions in the "Om Oss" section
This commit updates the links and accordions in the "Om Oss" section of the website. It replaces the link "/om-oss/hvordan-selge" with "/om-oss/selvregistrering" in the "HowToSell_Summary" component. Additionally, it updates the link "/om-oss/personlig-salgsprofil" to "/om-oss/salgsprofil" in the "AboutUsAccordions" component. These changes improve the navigation and user experience in the "Om Oss" section.